+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_list_feature_view_online_features_request.py
@@ -9,8 +9,10 @@ from darabonba.model import DaraModel
 class ListFeatureViewOnlineFeaturesRequest(DaraModel):
     def __init__(
         self,
+        config: str = None,
         join_ids: List[str] = None,
     ):
+        self.config = config
         # This parameter is required.
         self.join_ids = join_ids
 
@@ -22,6 +24,9 @@ class ListFeatureViewOnlineFeaturesRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.config is not None:
+            result['Config'] = self.config
+
         if self.join_ids is not None:
             result['JoinIds'] = self.join_ids
 
@@ -29,6 +34,9 @@ class ListFeatureViewOnlineFeaturesRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Config') is not None:
+            self.config = m.get('Config')
+
         if m.get('JoinIds') is not None:
             self.join_ids = m.get('JoinIds')