+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_list_feature_view_online_features_shrink_request.py
@@ -7,8 +7,10 @@ from darabonba.model import DaraModel
 class ListFeatureViewOnlineFeaturesShrinkRequest(DaraModel):
     def __init__(
         self,
+        config: str = None,
         join_ids_shrink: str = None,
     ):
+        self.config = config
         # This parameter is required.
         self.join_ids_shrink = join_ids_shrink
 
@@ -20,6 +22,9 @@ class ListFeatureViewOnlineFeaturesShrinkRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.config is not None:
+            result['Config'] = self.config
+
         if self.join_ids_shrink is not None:
             result['JoinIds'] = self.join_ids_shrink
 
@@ -27,6 +32,9 @@ class ListFeatureViewOnlineFeaturesShrinkRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Config') is not None:
+            self.config = m.get('Config')
+
         if m.get('JoinIds') is not None:
             self.join_ids_shrink = m.get('JoinIds')